✅ test(server): split high-water tests into PlaybackPositionHighWaterTest
